+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `isSliding` prop to `SingleEntryTable`.
 - Fix `ArrowIcon` display bug in Safari.
 
 ## [2.93.0] - 2020-09-30